refactor: extract text editor formatting actions
